requirements.txt
# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Medical Imaging
nibabel>=5.0.0
pydicom>=2.3.0
SimpleITK>=2.2.0

# Computer Vision & Image Processing
opencv-python>=4.7.0
albumentations>=1.3.0
Pillow>=9.5.0

# Machine Learning
scikit-learn>=1.2.0
scikit-image>=0.20.0

# Numerical Computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Experiment Tracking & Logging
tensorboard>=2.12.0
wandb>=0.15.0  # Optional: Weights & Biases
mlflow>=2.3.0  # Optional: MLflow

# Progress Bars
tqdm>=4.65.0

# Configuration
PyYAML>=6.0
omegaconf>=2.3.0

# Interpretability
grad-cam>=1.4.0
captum>=0.6.0  # PyTorch interpretability

# Testing
pytest>=7.3.0
pytest-cov>=4.0.0

# Code Quality
black>=23.3.0
flake8>=6.0.0
mypy>=1.2.0
isort>=5.12.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.22.0
ipywidgets>=8.0.0

# Utilities
python-dotenv>=1.0.0
argparse>=1.4.0

# Medical AI Specific
monai>=1.2.0  # Medical Open Network for AI (optional but useful)

# GPU Utilities (optional)
nvidia-ml-py3>=7.352.0  # For GPU monitoring
